+++ internal/pkg/conn/p2p/network_native.go
@@ -204,27 +204,6 @@ func NewNode(ctx context.Context, nodename string, nodeopt *options.NodeOptions,
 	return newnode, nil
 }
 
-func (node *Node) SetRumExchange(ctx context.Context, dbmgr *storage.DbMgr) {
-	peerStatus := NewPeerStatus()
-	var rexnotification chan RexNotification
-	rexnotification = make(chan RexNotification, 1)
-	var rexservice *RexService
-	rexservice = NewRexService(node.Host, peerStatus, node.NetworkName, ProtocolPrefix, rexnotification)
-	rexservice.SetDelegate()
-	rexchaindata := NewRexChainData(rexservice)
-	rexrelay := NewRexRelay(rexservice, dbmgr)
-	rexservice.SetHandlerMatchMsgType("rumchaindata", rexchaindata.Handler)
-	rexservice.SetHandlerMatchMsgType("rumrelay", rexrelay.Handler)
-	networklog.Infof("Enable protocol RumExchange")
-
-	node.peerStatus = peerStatus
-	node.RumExchange = rexservice
-
-	if rexnotification != nil {
-		go node.rexhandler(ctx, rexnotification)
-	}
-}
-
 func (node *Node) Bootstrap(ctx context.Context, config cli.Config) error {
 	var wg sync.WaitGroup
 	for _, peerAddr := range config.BootstrapPeers {
